requirements.txt
Flask
gunicorn
Flask-Cors
qrcode
Pillow